Refactored all the code from the manual provider.
Implement a new common interface to abstract multiple ways of
provisioning a machine.This way the code is more decoupled and making
the process easy to add more Provisioners
