+++ environs/manual/common/export_test.go
@@ -1,0 +1,8 @@
+// Copyright 2014 Canonical Ltd.
+// Licensed under the AGPLv3, see LICENCE file for details.
+
+package common
+
+var (
+	NetLookupHost = &netLookupHost
+)
