+++ environs/manual/linux/export_test.go
@@ -1,10 +1,9 @@
 // Copyright 2014 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package manual
+package linux
 
 var (
-	NetLookupHost         = &netLookupHost
 	ProvisionMachineAgent = &provisionMachineAgent
 )
