+++ environs/manual/linux/fakessh_test.go
@@ -1,7 +1,7 @@
 // Copyright 2013 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package manual_test
+package linux_test
 
 import (
 	"fmt"
@@ -13,7 +13,7 @@ import (
 	jc "github.com/juju/testing/checkers"
 	gc "gopkg.in/check.v1"
 
-	"github.com/juju/juju/environs/manual"
+	"github.com/juju/juju/environs/manual/linux"
 	"github.com/juju/juju/service"
 )
 
@@ -99,7 +99,7 @@ func installDetectionFakeSSH(c *gc.C, series, arch string) testing.Restorer {
 		"MemTotal: 4096 kB",
 		"processor: 0",
 	}, "\n")
-	return installFakeSSH(c, manual.DetectionScript, detectionoutput, 0)
+	return installFakeSSH(c, linux.DetectionScript, detectionoutput, 0)
 }
 
 // fakeSSH wraps the invocation of InstallFakeSSH based on the parameters.
