+++ environs/manual/linux/init_test.go
@@ -1,7 +1,7 @@
 // Copyright 2013 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package manual_test
+package linux_test
 
 import (
 	"strings"
@@ -9,7 +9,7 @@ import (
 	jc "github.com/juju/testing/checkers"
 	gc "gopkg.in/check.v1"
 
-	"github.com/juju/juju/environs/manual"
+	"github.com/juju/juju/environs/manual/linux"
 	"github.com/juju/juju/service"
 	"github.com/juju/juju/testing"
 )
@@ -27,8 +27,8 @@ func (s *initialisationSuite) TestDetectSeries(c *gc.C) {
 		"MemTotal: 4096 kB",
 		"processor: 0",
 	}, "\n")
-	defer installFakeSSH(c, manual.DetectionScript, response, 0)()
-	_, series, err := manual.DetectSeriesAndHardwareCharacteristics("whatever")
+	defer installFakeSSH(c, linux.DetectionScript, response, 0)()
+	_, series, err := linux.DetectSeriesAndHardwareCharacteristics("whatever")
 	c.Assert(err, jc.ErrorIsNil)
 	c.Assert(series, gc.Equals, "edgy")
 }
@@ -42,12 +42,12 @@ func (s *initialisationSuite) TestDetectionError(c *gc.C) {
 	}, "\n")
 	// if the script fails for whatever reason, then checkProvisioned
 	// will return an error. stderr will be included in the error message.
-	defer installFakeSSH(c, manual.DetectionScript, []string{scriptResponse, "oh noes"}, 33)()
-	hc, _, err := manual.DetectSeriesAndHardwareCharacteristics("hostname")
+	defer installFakeSSH(c, linux.DetectionScript, []string{scriptResponse, "oh noes"}, 33)()
+	hc, _, err := linux.DetectSeriesAndHardwareCharacteristics("hostname")
 	c.Assert(err, gc.ErrorMatches, "subprocess encountered error code 33 \\(oh noes\\)")
 	// if the script doesn't fail, stderr is simply ignored.
-	defer installFakeSSH(c, manual.DetectionScript, []string{scriptResponse, "non-empty-stderr"}, 0)()
-	hc, _, err = manual.DetectSeriesAndHardwareCharacteristics("hostname")
+	defer installFakeSSH(c, linux.DetectionScript, []string{scriptResponse, "non-empty-stderr"}, 0)()
+	hc, _, err = linux.DetectSeriesAndHardwareCharacteristics("hostname")
 	c.Assert(err, jc.ErrorIsNil)
 	c.Assert(hc.String(), gc.Equals, "arch=armhf cpu-cores=1 mem=4M")
 }
@@ -107,8 +107,8 @@ func (s *initialisationSuite) TestDetectHardwareCharacteristics(c *gc.C) {
 	for i, test := range tests {
 		c.Logf("test %d: %s", i, test.summary)
 		scriptResponse := strings.Join(test.scriptResponse, "\n")
-		defer installFakeSSH(c, manual.DetectionScript, scriptResponse, 0)()
-		hc, _, err := manual.DetectSeriesAndHardwareCharacteristics("hostname")
+		defer installFakeSSH(c, linux.DetectionScript, scriptResponse, 0)()
+		hc, _, err := linux.DetectSeriesAndHardwareCharacteristics("hostname")
 		c.Assert(err, jc.ErrorIsNil)
 		c.Assert(hc.String(), gc.Equals, test.expectedHc)
 	}
@@ -117,43 +117,43 @@ func (s *initialisationSuite) TestDetectHardwareCharacteristics(c *gc.C) {
 func (s *initialisationSuite) TestCheckProvisioned(c *gc.C) {
 	listCmd := service.ListServicesScript()
 	defer installFakeSSH(c, listCmd, "", 0)()
-	provisioned, err := manual.CheckProvisioned("example.com")
+	provisioned, err := linux.CheckProvisioned("example.com")
 	c.Assert(err, jc.ErrorIsNil)
 	c.Assert(provisioned, jc.IsFalse)
 
 	defer installFakeSSH(c, listCmd, "juju...", 0)()
-	provisioned, err = manual.CheckProvisioned("example.com")
+	provisioned, err = linux.CheckProvisioned("example.com")
 	c.Assert(err, jc.ErrorIsNil)
 	c.Assert(provisioned, jc.IsTrue)
 
 	// stderr should not affect result.
 	defer installFakeSSH(c, listCmd, []string{"", "non-empty-stderr"}, 0)()
-	provisioned, err = manual.CheckProvisioned("example.com")
+	provisioned, err = linux.CheckProvisioned("example.com")
 	c.Assert(err, jc.ErrorIsNil)
 	c.Assert(provisioned, jc.IsFalse)
 
 	// if the script fails for whatever reason, then checkProvisioned
 	// will return an error. stderr will be included in the error message.
 	defer installFakeSSH(c, listCmd, []string{"non-empty-stdout", "non-empty-stderr"}, 255)()
-	_, err = manual.CheckProvisioned("example.com")
+	_, err = linux.CheckProvisioned("example.com")
 	c.Assert(err, gc.ErrorMatches, "subprocess encountered error code 255 \\(non-empty-stderr\\)")
 }
 
 func (s *initialisationSuite) TestInitUbuntuUserNonExisting(c *gc.C) {
 	defer installFakeSSH(c, "", "", 0)() // successful creation of ubuntu user
 	defer installFakeSSH(c, "", "", 1)() // simulate failure of ubuntu@ login
-	err := manual.InitUbuntuUser("testhost", "testuser", "", nil, nil)
+	err := linux.InitUbuntuUser("testhost", "testuser", "", nil, nil)
 	c.Assert(err, jc.ErrorIsNil)
 }
 
 func (s *initialisationSuite) TestInitUbuntuUserExisting(c *gc.C) {
 	defer installFakeSSH(c, "", nil, 0)()
-	manual.InitUbuntuUser("testhost", "testuser", "", nil, nil)
+	linux.InitUbuntuUser("testhost", "testuser", "", nil, nil)
 }
 
 func (s *initialisationSuite) TestInitUbuntuUserError(c *gc.C) {
 	defer installFakeSSH(c, "", []string{"", "failed to create ubuntu user"}, 123)()
 	defer installFakeSSH(c, "", "", 1)() // simulate failure of ubuntu@ login
-	err := manual.InitUbuntuUser("testhost", "testuser", "", nil, nil)
+	err := linux.InitUbuntuUser("testhost", "testuser", "", nil, nil)
 	c.Assert(err, gc.ErrorMatches, "subprocess encountered error code 123 \\(failed to create ubuntu user\\)")
 }
