+++ environs/manual/linux/suite_test.go
@@ -1,7 +1,7 @@
 // Copyright 2013 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package manual_test
+package linux_test
 
 import (
 	"runtime"
